pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artisanflow"
version = "0.1.0"
description = "Automated artisan craft fulfillment and marketplace platform for Oaxaca, Mexico"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "ArtisanFlow Team"}]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "anthropic>=0.39.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "duckduckgo-search>=4.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiofiles>=23.0.0",
]

[project.scripts]
artisanflow = "artisanflow.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
